malloc and free

+++ 0x0B-malloc_free/2-str_concat.c
@@ -7,7 +7,7 @@
  * Return: string length
  */
 
-int _strlen(char *s);
+size_t _strlen(char *s);
 {
 	size_t len;
